Good parameters and preparing for object oriented version of code
Phase space tunnels still not finished.
Diagonals of synaptic matrix weights set to 0
Parameters changed to values that, for 500 neurons, provides a somewhat variable time-dependent response to constant stimulus. Also chosen such that the network response disappears when the stimulus does.
To be used as a basis for object version of the code which allows easy duplication of the network, specification of parameters (much like the demo code), and linking between networks.